Fixes after review part 2
