+++ src/PrestaShopBundle/Resources/config/routing/admin/configure/advanced_parameters/employee.yml
@@ -71,7 +71,6 @@ admin_employees_edit:
   requirements:
     employeeId: \d+
 
-
 admin_employees_toggle_navigation:
   path: /toggle-navigation
   methods: [POST]
