+++ src/PrestaShopBundle/Resources/config/services/core/form/form_data_provider.yml
@@ -79,4 +79,4 @@ services:
       - '@prestashop.core.query_bus'
       - '@=service("prestashop.adapter.multistore_feature").isActive()'
       - '@=service("prestashop.adapter.shop.context").getShops(true, true)'
-      - '@=service("prestashop.adapter.employee.default_avatar_provider").getDefaultAvatarUrl()'
+      - '@=service("prestashop.adapter.employee.avatar_provider").getDefaultAvatarUrl()'
